Rename EntityRotationPolicy to EntityRotation
