+++ common/CMakeLists.txt
@@ -121,7 +121,7 @@ set(COMMON_SOURCE
         ${COMMON_SOURCE_DIR}/Model/EntityNodeIndex.cpp
         ${COMMON_SOURCE_DIR}/Model/EntityProperties.cpp
         ${COMMON_SOURCE_DIR}/Model/EntityPropertiesVariableStore.cpp
-        ${COMMON_SOURCE_DIR}/Model/EntityRotationPolicy.cpp
+        ${COMMON_SOURCE_DIR}/Model/EntityRotation.cpp
         ${COMMON_SOURCE_DIR}/Model/Game.cpp
         ${COMMON_SOURCE_DIR}/Model/GameConfig.cpp
         ${COMMON_SOURCE_DIR}/Model/GameEngineConfig.cpp
@@ -611,7 +611,7 @@ set(COMMON_HEADER
         ${COMMON_SOURCE_DIR}/Model/EntityNodeIndex.h
         ${COMMON_SOURCE_DIR}/Model/EntityProperties.h
         ${COMMON_SOURCE_DIR}/Model/EntityPropertiesVariableStore.h
-        ${COMMON_SOURCE_DIR}/Model/EntityRotationPolicy.h
+        ${COMMON_SOURCE_DIR}/Model/EntityRotation.h
         ${COMMON_SOURCE_DIR}/Model/Game.h
         ${COMMON_SOURCE_DIR}/Model/GameConfig.h
         ${COMMON_SOURCE_DIR}/Model/GameEngineConfig.h
