+++ common/src/Model/EntityRotation.h
